Tensorflow/Keras adapters refactoring
