+++ aim/sdk/adapters/keras_mixins.py
@@ -1,7 +1,3 @@
-from typing import Optional, List
-from aim.sdk.run import Run
-
-
 class TrackerKerasCallbackMetricsEpochEndMixin(object):
     def on_epoch_end(self, epoch, logs=None):
         # Log metrics
@@ -33,34 +29,3 @@ class TrackerKerasCallbackMetricsEpochEndMixin(object):
         lr = self._get_learning_rate()
         if lr is not None:
             track_func(lr, name='lr', epoch=epoch, context={'subset': 'train'})
-
-
-def get_keras_tracker_callback(keras_callback_cls, mixins: List):
-    class KerasTrackerCallback(keras_callback_cls, *mixins):
-        def __init__(self, repo: Optional[str] = None,
-                     experiment: Optional[str] = None,
-                     run: Optional[Run] = None):
-            super(KerasTrackerCallback, self).__init__()
-
-            if run is None:
-                if repo is None and experiment is None:
-                    self._run = Run()
-                else:
-                    self._run = Run(repo=repo, experiment=experiment)
-            else:
-                print('Passing Run instance to AimCallback will be '
-                      'deprecated in future versions, '
-                      'pass the callback arguments explicitly')
-                self._run = run
-
-        @property
-        def run(self) -> Run:
-            return self._run
-
-        def on_epoch_end(self, *args, **kwargs):
-            for mixin_cls in mixins:
-                if 'on_epoch_end' in mixin_cls.__dict__:
-                    mixin_cls.on_epoch_end(self, *args, **kwargs)
-                    return
-
-    return KerasTrackerCallback
